Refactor: Split ScriptObjectStorage into its own file
